API

PileResponse.find_cpt_column_names — Function

find_cpt_column_names(names::Vector{SubString{String}})

Returns a vector `[depth_col, qc_col, fs_col, u2_col]` providing the relevant names for each item from `names`.

The function looks for the strings "depth", "qc", "fs" and "u2" within the items in names. It is not case sensitive.

Example

data = read_delimited_text_file("./data/example_cpt_data.csv", delim=',')
column_names = collect(keys(data))
depth_col, qc_col, fs_col, u2_col = find_cpt_column_names(column_names)

PileResponse.get_least_squares_interpolator — Function

get_least_squares_interpolator(xvals::AbstractVector, yvals::AbstractVector; p0::AbstractVector = [0.0,0.0])

Returns a function `fun()` such that `fun(x) = mx + b` is the least squares linear approximation to [xvals, yvals]

p0 [Float64, Float64] is the initial guess, and [0.0, 0.0] seems to work

PileResponse.show_table — Function

function show_table(columns::AbstractArray, header::AbstractVector{String}; num_rows::Int64=5, printformat="%5.3f")

Show a formatted table in HTML with given column data and headers.

columns = [col1data, col2data, ..] a list of vectors with the data for each column
header = ["col1 name", "col2 name", ...] a list of strings with the column names

PileResponse.get_pile_head_displacement — Function

get_pile_head_displacement(k0::Float64, pile_head_loads::AbstractVector{Float64}, pile_ult_resistance::Float64)

Returns the pile head displacement for the given load vector

The assumption is that the pile head stiffness k for a given load P can be approximated as

$k = k_{0}(1 - (P/P_{ult})^{0.3})$

For further details see:

Mayne, P. W. (2001) Stress-strain-strength flow parameters from enhanced in-situ tests.

Fahey, M. and Carter, J. P. (1993) A finite element study of the pressuremeter in sand using a nonlinear elastic plastic model.

k0 is the initial, or small strain, pile head stiffness
pile_head_loads is a vector of the load applied at the pile head. It will be sorted into increasing order.
pile_ult_resistance is the assessed ultimate load capacity of the pile

PileResponse.get_initial_pile_head_stiffness — Function

get_initial_pile_head_stiffness(pile_length::Float64, pile_diameter::Float64, Epile::Int64, Esoil_L::Float64, Esoil_Lon2::Float64; ν::Float64 = 0.3)

Returns the initial, or small strain, pile head stiffness (MN/m)

The theory is based on the closed form elastic solution provided by Randolph and Wroth (1978), Analysis of deformation of vertically loaded piles.

The theory assumes that the soil has a linearly increasing elastic modulus with depth.

Epile (MPa) is the elastic modulus of the pile
Esoil_L (MPa) is the small strain (E₀) elastic modulus of the soil at the base of the pile shaft
Esoil_Lon2 (MPa) is the small strain (E₀) elastic modulus of the soil at the midpoint of the pile shaft
ν (input as \nu[tab]) is the Poisson's ratio of the soil

PileResponse.get_soil_type_CPT2012 — Function

get_soil_type_CPT2012(Ic::AbstractVector{Float64})

Returns vector of soil types for use in CPT2012 capacity assessment:
	1 = silts and clays (Ic = 2.60 - 3.0)
	2 = intermediate soil (Ic = 2.05 - 2.60)
	3 = sands (Ic = 0.0 - 2.05)

PileResponse.get_kc_base_CPT2012 — Function

get_kc_base_CPT2012()

Returns a dictionary with pile base resistance kc factors such that Qult_base = kc * qca, where qca is the equivalent average cone resistance at base (see: get_average_qc_at_pile_base()).

Keys: pile types
Values: kc factors for [silt/clay, intermediate soils, sands]

Reference: Frank. R. (2017) Some aspects of pile design in France

PileResponse.get_ultimate_shaft_load — Function

get_ultimate_shaft_load(depth_m::AbstractVector{Float64}, fshaft_MPa::AbstractVector{Float64}, pile_diameter::Float64, pile_toe_depth::Float64)

returns the ultimate shaft load for the pile (MN)

fshaft_MPa is the ultimate shaft resistance (MPa) for each node corresponding to the depth_m vector.

PileResponse.get_ultimate_shaft_resistance — Function

get_ultimate_shaft_resistance(qc_MPa::AbstractVector{Float64}, Ic::AbstractVector{Float64}, pile_type::String; factor::Float64 = 1.0)

Returns the shaft resistance (MPa) for each element.

Set factor = 1.0 for ultimate resistance, and factor < 1.0 for resistances below the ultimate load

PileResponse.get_fsmax_shaft_CPT2012 — Function

get_fsmax_shaft_CPT2012()

Returns a dictionary with maximum pile shaft resistance values fsmax,
	such that `Qult_shaft <= fsmax`.

Keys: pile types
Values: maximum shaft resistance fsmax(kPa) for [silt/clay, intermediate soils, sands]

Reference: Frank. R. (2017) Some aspects of pile design in France

PileResponse.get_alpha_shaft_CPT2012 — Function

get_alpha_shaft_CPT2012()

Returns a dictionary with pile shaft resistance α factors,
	such that Qult_shaft = α * fsol (see: get_fsol_shaft_CPT2012()).

Keys: pile types
Values: alpha factors for [silt/clay, intermediate soils, sands]

Reference: Frank. R. (2017) Some aspects of pile design in France

PileResponse.get_fsol_shaft_CPT2012 — Function

get_fsol_shaft_CPT2012(qc_MPa::AbstractVector{Float64}, Ic::AbstractVector{Float64})

Returns fsol(kPa)::Vector the unfactored ultimate pile shaft resistance.

Ic is CPT soil behaviour type index

Reference: Frank. R. (2017) Some aspects of pile design in France

PileResponse.get_average_qc_at_pile_base — Function

get_average_qc_at_pile_base(depth_m::AbstractVector{Float64},qc_MPa::AbstractVector{Float64}, pile_toe_depth::Float64, pile_diameter::Float64; clip_to_30pct::Bool = false)

Returns average qc within +/- 1.5 pile diameters from the toe, with values limited to within +/- 30% of the value at the toe

if clip_to_30pct = false, values will not be limited to +/- 30% of the value at the toe prior to averaging

PileResponse.get_E0 — Function

get_E0(Vs::AbstractVector{Float64}; gamma::Float64 = 18.0, ν::Float64 = 0.3)

returns `E₀(MPa) = 2 * (1 + ν) * gamma / 9.81 * Vs² * 0.001`

gamma is soil unit weight in kN/m², assumed constant over depth_m
ν is the Poisson's ratio

PileResponse.get_Vs — Function

get_Vs(depth_m::AbstractVector{Float64}, qc_MPa::AbstractVector{Float64}, fs_MPa::AbstractVector{Float64}, u2_MPa::AbstractVector{Float64}, gw_depth::Float64; gamma::Float64 = 18.0, a::Float64 = 0.73)

returns `(αᵥₛ * qn / pa)^0.5`
	where `αᵥₛ = 10^(0.55 * Ic + 1.68)`

gw_depth is depth to groundwater in metres
gamma is soil unit weight in kN/m², assumed constant over depth_m
a is the net area ratio of the cone, typically between 0.70 and 0.85

PileResponse.get_Ic — Function

get_Ic(depth_m::AbstractVector{Float64}, qc_MPa::AbstractVector{Float64},
fs_MPa::AbstractVector{Float64}, u2_MPa::AbstractVector{Float64}, gw_depth::Float64; gamma::Float64 = 18.0, a::Float64 = 0.73)\
returns `[(3.47 - log10(Qt))² + (log10(Fr) + 1.22)²]^0.5`

gw_depth is depth to groundwater in metres
gamma is soil unit weight in kN/m², assumed constant over depth_m
a is the net area ratio of the cone, typically between 0.70 and 0.85

PileResponse.get_bigQt — Function

get_bigQt(depth_m::AbstractVector{Float64}, qc_MPa::AbstractVector{Float64}, u2_MPa::AbstractVector{Float64}; gw_depth::Float64, gamma::Float64 = 18.0, a::Float64 = 0.73)\
returns `Qₜ = (qₜ - σᵥ₀) / σ'ᵥ₀`\

gw_depth is depth to groundwater in metres
gamma is soil unit weight in kN/m², assumed constant over depth_m
a is the net area ratio of the cone, typically between 0.70 and 0.85

PileResponse.get_Fr — Function

function get_Fr(depth_m::AbstractVector{Float64},  qc_MPa::AbstractVector{Float64}, fs_MPa::AbstractVector{Float64},u2_MPa::AbstractVector{Float64}; gamma::Float64 = 18.0, a::Float64 = 0.73)\
returns `(fₛ / qₙ) * 100`

a is the net area ratio of the cone, typically between 0.70 and 0.85 gamma is soil unit weight in kN/m², assumed constant over depth_m

PileResponse.get_Rf — Function

get_Rf(qc_MPa::AbstractVector{Float64}, fs_MPa::AbstractVector{Float64}, u2_MPa::AbstractVector{Float64}; a::Float64 = 0.73)\
Returns `(fₛ / qₜ) * 100`

a is the net area ratio of the cone, typically between 0.70 and 0.85

PileResponse.get_qn — Function

function get_qn(depth_m::AbstractVector{Float64},  qc_MPa::AbstractVector{Float64}, u2_MPa::AbstractVector{Float64}; gamma::Float64 = 18.0, a::Float64 = 0.73)\
returns qₙ = qt - σᵥ₀

a is the net area ratio of the cone, typically between 0.70 and 0.85 gamma is soil unit weight in kN/m², assumed constant over depth_m

PileResponse.get_qt — Function

get_qt(qc_MPa::AbstractVector{Float64}, u2_MPa::AbstractVector{Float64}; a::Float64 = 0.73)\
Returns `qt = qc + u2(1 - a)`

where a is the net area ratio of the cone, typically between 0.70 and 0.85

PileResponse.get_sigmav0_effective — Function

get_sigmav0_effective(depth_m::AbstractVector{Float64}, gw_depth::Float64, gamma::Float64 = 18.0)\
Returns the effective vertical stress σ'ᵥ₀(MPa)

gw_depth is depth to groundwater in metres
gamma is soil unit weight in kN/m², assumed constant over depth_m

PileResponse.get_sigmav0_total — Function

get_sigmav0_total(depth_m::AbstractVector{Float64}; gamma::Float64 = 18.0)\
Returns the total vertical stress σᵥ₀(MPa)

gamma is unit weight in kN/m²

PileResponse.read_delimited_text_file — Function

read_delimited_text_file(filepath::String; delim::AbstractChar=',', T::Type=Float64)\
Returns a dictionary `data` such that `data[col1_header] = col1_values`

Read a delimited text file. Assumes the first row is the header and that all columns below the header have the same type.